% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_percentages.R
\name{calc_percentage_gewogen}
\alias{calc_percentage_gewogen}
\title{FUNCTIE: calc_prevalentie_gewogen()}
\usage{
calc_percentage_gewogen(var, dsgn)
}
\arguments{
\item{var}{naam van de variabele (string).}

\item{dsgn}{een svydesign object (met gewichten).}
}
\value{
weigthed precentages
}
\description{
Doel: Gewogen aantallen en percentages berekenen
voor een categorische variabele binnen een survey-design.
}
\examples{
data("df_example")
## Gebruik je survey-design zoals het is
design <- svydesign(
 id      = ~RESPNR,          # of ID
 weights = ~WT1,             # jouw gewichtvariabele
 data    = df_example        # jouw dataset
)
## Gebruik voor één variabele:
calc_prevalentie_gewogen("yypandi_Ax1", design)

}
