% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unweighted_percentages.R
\name{calc_percentage_ongewogen}
\alias{calc_percentage_ongewogen}
\title{FUNCTIE: calc_percentage_ongewogen()}
\usage{
calc_percentage_ongewogen(vars, data, include_na = FALSE)
}
\arguments{
\item{vars}{variabelenaam of vector met variabelen}

\item{data}{is dataset}

\item{include_na}{is TRUE/FALSE (missings meenemen als categorie)}
}
\value{
unweigthed precentages
}
\description{
Doel:Ongewogen aantallen en percentages berekenen voor één of meerdere categorische variabelen.
}
\examples{
Hier moet nog een voorbeeld komen
}
