% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{log_regressie_gewogen}
\alias{log_regressie_gewogen}
\title{FUNCTIE: log_regressie_gewogen()}
\usage{
log_regressie_gewogen(dep, ind, correct = NULL, design)
}
\arguments{
\item{dep}{vector met afhankelijke variabelen (binaire factoren, 0/1)}

\item{ind}{categorische voorspeller (factor met ≥ 2 niveaus)}

\item{correct}{optionele vector met controlevariabelen (bv. leeftijd, opleiding)}

\item{design}{svydesign-object (survey design met gewichten)}
}
\value{
Tabel met per uitkomstmaat en per vergelijking
}
\description{
De functie is bedoeld om gewogen logistische regressies uit te voeren voor één of meerdere binaire uitkomstmaten, met een categorische voorspeller met twee of meer niveaus. Het uitvoeren van gewogen logistische regressies zonder interacties voor meerdere binaire uitkomstmaten. Voor elke uitkomstmaat worden alle unieke paarsgewijze vergelijkingen tussen de niveaus van een categorische onafhankelijke variabele uitgevoerd (bijv. bij 3 niveaus: 3 vergelijkingen).
}
\examples{
data("df_log_weighted")
library(survey)
library(trimbosR)
Survey-design instellen
design <- svydesign(
ids = ~xgemnr,
strata = ~xlandsdeel,
weights = ~WT1,
data = df_log_weighted,
nest = TRUE
)
###  Gebruik voor één of meerdere uitkomstmaten:
log_regressie_gewogen(
dep = c("geslacht"),                # afhankelijke vars
ind = "leeftijd",                   # onafhankelijke var
correct = "xgender",                # covariaten
design = design                     # survey-design
)

}
