% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{log_regressie_gewogen}
\alias{log_regressie_gewogen}
\title{FUNCTIE: log_regressie_gewogen()}
\usage{
log_regressie_gewogen(dep, ind, correct = NULL, design)
}
\arguments{
\item{dep}{vector met afhankelijke variabelen (binaire factoren, 0/1)}

\item{ind}{categorische voorspeller (factor met ≥ 2 niveaus)}

\item{correct}{optionele vector met controlevariabelen (bv. leeftijd, opleiding)}

\item{design}{svydesign-object (survey design met gewichten)}
}
\value{
Tabel met per uitkomstmaat en per vergelijking
}
\description{
De functie is bedoeld om gewogen logistische regressies uit te voeren voor één of meerdere binaire uitkomstmaten, met een categorische voorspeller met twee of meer niveaus. Het uitvoeren van gewogen logistische regressies zonder interacties voor meerdere binaire uitkomstmaten. Voor elke uitkomstmaat worden alle unieke paarsgewijze vergelijkingen tussen de niveaus van een categorische onafhankelijke variabele uitgevoerd (bijv. bij 3 niveaus: 3 vergelijkingen).
}
\examples{
library(trimbosR)
data("df_example")
library(survey)

df_example$Marijuana <- ifelse(df_example$Marijuana == "Yes", 1,
ifelse(df_example$Marijuana == "No", 0, NA))  ## Change Yes/No to 1 - 0 (still includes NAs)
df_example$SmokeNow <- ifelse(df_example$SmokeNow == "Yes", 1,
ifelse(df_example$SmokeNow == "No", 0, NA))  ## Change Yes/No to 1 - 0 (still includes NAs)
vars <- c("ID", "SmokeNow", "Marijuana", "Age", "MaritalStatus", "Gender") ## Optional: select complete cases for the model
df_example <- df_example[complete.cases(df_example[, vars]), ] ## Select complete cases
## Survey-design instellen
design <- survey::svydesign(
ids = ~ID,
strata = ~Gender,
weights = ~WT1,
data = df_example,
nest = TRUE
)
###  Gebruik voor één of meerdere uitkomstmaten:
log_regressie_gewogen(
dep = c("Marijuana", "SmokeNow"),   # afhankelijke vars
ind = "Age",                        # onafhankelijke var
correct = "MaritalStatus",          # covariaten
design = design                     # survey-design
)

}
